renames filename runRecheck to runRetryWhen
